refactor: extract NodeEvent utils into separate module
