+++ app/screens/node/Node.tsx
@@ -8,7 +8,6 @@ import {
   WrapperWith2SideBars,
   Button,
   Link,
-  CustomTimeAgo,
   ColorStatusIndicator,
 } from '../../basicComponents';
 import { hideSmesherLeftPanel, setUiError } from '../../redux/ui/actions';
@@ -29,7 +28,6 @@ import {
   NodeStatus,
   PostSetupState,
   RewardsInfo,
-  NodeEvent,
 } from '../../../shared/types';
 import { isWalletOnly } from '../../redux/wallet/selectors';
 import * as SmesherSelectors from '../../redux/smesher/selectors';
@@ -46,7 +44,8 @@ import {
   epochByLayer,
   nextEpochTime,
 } from '../../../shared/layerUtils';
-import { convertBytesToMiB, getEventType } from '../../../shared/utils';
+import { convertBytesToMiB } from '../../../shared/utils';
+import NodeEventActivityRow from './NodeEventActivityRow';
 
 const Wrapper = styled.div`
   display: flex;
@@ -298,97 +297,6 @@ const ERR_MESSAGE_ERR_STATE =
 const ERR_MESSAGE_NODE_ERROR =
   'The Node is not syncing. Please check the Network tab';
 
-const withTime = (str: string, time: number) => (
-  <>
-    {str}{' '}
-    <CustomTimeAgo
-      time={time}
-      dict={{
-        prefixAgo: 'in',
-        prefixFromNow: 'in',
-        suffixAgo: 'ago',
-        suffixFromNow: null,
-        seconds: '%d seconds',
-      }}
-    />
-  </>
-);
-
-const getStageName = (event: NodeEvent) => {
-  switch (getEventType(event)) {
-    case 'initStart':
-      return 'PoST data initialization';
-    case 'initComplete':
-      return 'PoST data initialization complete';
-    case 'poetWaitRound':
-      return 'Waiting for PoET registration';
-    case 'poetWaitProof':
-      return 'Waiting for PoET proof';
-    case 'postStart':
-      return 'PoST proof generation';
-    case 'postComplete':
-      return 'Generating PoST proof complete';
-    case 'atxPublished':
-      return 'Publishing activation';
-    case 'eligibilities':
-      return 'Calculating eleigibilities';
-    case 'proposal':
-      return 'Publishing proposal';
-    case 'beacon':
-      return 'Generating beacon';
-    default:
-      return `Unknown "${getEventType(event)}"`;
-  }
-};
-
-const renderNodeActivity = (event: NodeEvent) => {
-  if (event && event?.failure) {
-    return (
-      <ErrorMessage>
-        Stage &quot;{getStageName(event)}&quot; failed. Check the logs for more
-        details.
-      </ErrorMessage>
-    );
-  }
-  switch (getEventType(event)) {
-    case 'initStart':
-      return 'Started PoST data initialization';
-    case 'initComplete':
-      return 'Completed PoST data initialization';
-    case 'poetWaitRound':
-      return withTime(
-        'Waiting for PoET registration window',
-        event.timestamp + (event.poetWaitRound?.wait || 0)
-      );
-    case 'poetWaitProof': {
-      return withTime(
-        'Waiting for PoET challenge',
-        event.timestamp + (event.poetWaitProof?.wait || 0)
-      );
-    }
-    case 'postStart':
-      return "Generating PoST proof for the PoET's challenge";
-    case 'postComplete':
-      return 'Finished generating PoST proof';
-    case 'atxPublished':
-      return 'Published activation. Waiting for the next epoch';
-    case 'eligibilities':
-      return event?.eligibilities?.eligibilities
-        ? `Eligible for rewards in layers ${event.eligibilities.eligibilities
-            .map((el) => el.layer)
-            .join(', ')}`
-        : `Computed eligibilities for the epoch ${
-            event.eligibilities?.epoch || ''
-          }`;
-    case 'proposal':
-      return `Published proposal on layer ${event.proposal?.layer}`;
-    case 'beacon':
-      return `Node computed randomness beacon for epoch ${event.beacon?.epoch}`;
-    default:
-      return event?.help || 'Node is preparing...';
-  }
-};
-
 const SmesherStatus = ({
   smesherId,
   status,
@@ -574,7 +482,7 @@ const Node = ({ history, location }: Props) => {
             <ColorStatusIndicator
               color={lastEvent?.failure ? smColors.red : smColors.green}
             />
-            <EventText>{renderNodeActivity(lastEvent)}</EventText>
+            <EventText>{NodeEventActivityRow(lastEvent)}</EventText>
             <Link
               text="Open logs"
               onClick={() => history.push(MainPath.NodeEvents)}
